# Remove the complicated block search logic for a simpler branchless