binary search

The code is simpler and faster.

```
Before
test postings::bench::bench_segment_intersection                                                                         ... bench:   2,093,697 ns/iter (+/- 115,509)
test postings::bench::bench_skip_next_p01                                                                                ... bench:      58,585 ns/iter (+/- 796)
test postings::bench::bench_skip_next_p1                                                                                 ... bench:     160,872 ns/iter (+/- 5,164)
test postings::bench::bench_skip_next_p10                                                                                ... bench:     615,229 ns/iter (+/- 25,108)
test postings::bench::bench_skip_next_p90                                                                                ... bench:   1,120,509 ns/iter (+/- 22,271)
```

```
After
test postings::bench::bench_skip_next_p01                                                                                ... bench:      51,701 ns/iter (+/- 719)
test postings::bench::bench_skip_next_p1                                                                                 ... bench:     124,848 ns/iter (+/- 2,475)
test postings::bench::bench_skip_next_p10                                                                                ... bench:     453,206 ns/iter (+/- 7,315)
test postings::bench::bench_skip_next_p90                                                                                ... bench:     880,564 ns/iter (+/- 6,158)

```